tests/wpt/web-platform-tests/html/cross-origin-embedder-policy/require-corp-sw.https.html

<!doctype html>
<title>Cross Origin Embedder Policy: requests initiated from a service worker with 'require-corp'</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>
'use strict';

const SCRIPT = 'resources/require-corp-sw.js';
const SCOPE = 'resources/in-scope';
let worker = null;

promise_test(async t => {
  const registration =
      await service_worker_unregister_and_register(t, SCRIPT, SCOPE);
  promise_test(async t => registration.unregister(), 'Clean up global state');
  worker = registration.installing;
  await wait_for_state(t, worker, 'activated');
}, 'Set up global state');

promise_test(async t => {
  const p = new Promise(resolve =>
      navigator.serviceWorker.addEventListener('message', resolve,
                                               {once: true}));
  worker.postMessage('WithCorp');
  assert_equals((await p).data, 'opaque');
}, "fetch() to 'CORP: cross-origin' response should succeed.");

promise_test(async t => {
  const p = new Promise(resolve =>
      navigator.serviceWorker.addEventListener('message', resolve,
                                               {once: true}));
  worker.postMessage('WithoutCorp');
  assert_equals((await p).data, 'TypeError: Failed to fetch');
}, "fetch() to no CORP response should not succeed.");
</script>
